[CFDS] [Refactoring] Hamza / Remove the hard-coded variables and replace them with config file (#11575)
* refactor: split files into respected folders and remove hard-coded variables

* refactor: used aliases and added new Helper Alias
