+++ packages/cfd/tsconfig.json
@@ -7,6 +7,7 @@
             "Constants/*": ["./src/Constants/*"],
             "Components/*": ["./src/Components/*"],
             "Containers/*": ["./src/Containers/*"],
+            "Helpers/*": ["./src/Helpers/*"],
             "Modules/*": ["./src/Modules/*"],
             "Sass/*": ["./src/sass/*"],
             "Stores/*": ["./src/Stores/*"],
